+++ src/components/sections/orders.js
@@ -85,6 +85,9 @@ const Orders = () => {
   };
 
   const handleDelete = async (id) => {
+    const confirmDelete = window.confirm("¿Estás seguro de que deseas eliminar este pedido?");
+    if (!confirmDelete) return;
+
     try {
       await deleteOrder(id);
       setOrders(orders.filter(order => order.id !== id));
@@ -225,6 +228,9 @@ const Orders = () => {
   };
 
   const handleDeleteProduct = async (order, productId) => {
+    const confirmDelete = window.confirm("¿Estás seguro de que deseas eliminar este producto del pedido?");
+    if (!confirmDelete) return;
+
     try {
       const token = localStorage.getItem('token');
       await axios.delete(`https://vikingsdb.up.railway.app/products_order/${order.id}/${productId}`, {
